create DraggableTray optionally controlled component
